requirements.txt
fastapi
uvicorn
aioredis
pymysql
sqlalchemy
databases
pyjwt
python-dotenv
pydantic-settings
starlette
aiomysql
pycryptodome
bcrypt